Funcionalidade de criação de pedido Funcionalidade de atualização de pedido v1

+++ SalesService/SalesService.Domain/Aggregates/Orders/Item.cs
@@ -1,4 +1,4 @@
 ﻿namespace SalesService.Domain.Aggregates.Orders
 {
-    public record Item(Guid ProductId, decimal Units);
+    public record Item(Guid ProductId, int Units);
 }
